Move RxErrorsRule to test utils for sharing
